+++ resources/dev.gethop.rbac.next/rbac-tables.pg.down.sql
@@ -2,6 +2,8 @@ DROP TABLE IF EXISTS rbac_context_type CASCADE;
 -- ;;
 DROP TABLE IF EXISTS rbac_context CASCADE;
 -- ;;
+DROP TABLE IF EXISTS rbac_context_parent CASCADE;
+-- ;;
 DROP TABLE IF EXISTS rbac_role CASCADE;
 -- ;;
 DROP TABLE IF EXISTS rbac_role_assignment CASCADE;